requirements.txt
python-telegram-bot==20.3
aiohttp
apscheduler
reportlab
pillow
requests
